Add files via upload
This code is used to test which vesicles are being detected in individual images. Use this as a quick check to assess whether there is sufficient contrast between your vesicles and the background to detect the edges, and assign suitable values for the sensitivity_factor and min/max diameter variables.